pyproject.toml
[tool.poetry]
name = "auto_cronometer"
version = "0.1.0"
description = "Programmtically interact with Cronometer"
authors = ["Ben Yang <ben.yang@utexas.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^3.141.0"
google-api-python-client = "^1.12.8"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
docopt = "^0.6.2"
PyYAML = "^5.4.1"
httpx = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
autocm = "auto_cronometer.main:main"
